update specs name
